上課內容/學院_11202python應用實戰班/grid/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
Pillow==9.4.0
requests==2.28.2
urllib3==1.26.15
